requirements.txt
django
dj-rest-auth
djangorestframkework
django-allauth   
requests
jwt